Refactor eslint code to reuse with other linters
